Renamed the trackerDownload folder in the modules directory and updated all import paths where the folder is used.
Renamed the trackerDownload folder in the modules directory and updated all import paths where the folder is used.
