+++ modules/swagger-api/init.js
@@ -33,7 +33,7 @@ exports.init = (app, apiUrl) => {
             './modules/notification/prepare-notification-data/routes.js',
             "./modules/notification/notification-middleware/routes.js",
             "./modules/project-setting/routes.js",
-            "./modules/trackerDownload/routes.js",
+            "./modules/tracker-download/routes.js",
             "./modules/task-index/routes.js",
             "./modules/create-project/routes.js",
             "./modules/notification-count/routes.js",
